backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.1
torchvision==0.17.1
openai-whisper==20231117
ffmpeg-python==0.2.0
pydantic==2.6.3
opencv-python-headless
ultralytics
python-multipart
